# docs/api-spec.yaml
openapi: "3.0.3"
info:
  title: Crawler.io API
  description: Unified Social Media Data API — one integration, all platforms.
  version: "1.0.0"
  contact:
    email: api@crawler.io
    url: https://docs.crawler.io
servers:
  - url: https://api.crawler.io/v1
    description: Production
  - url: https://sandbox.crawler.io/v1
    description: Sandbox (mock data)
security:
  - BearerAuth: []
...
paths:
...
  /creators/{creator_id}/posts:
    get:
      tags: [Posts]
      summary: Get posts by creator
      parameters:
        - name: creator_id
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [image, carousel, video, reel, short, story, live]
        - name: sort
          in: query
          schema:
            type: string
            enum: ["-published_at", "published_at", "-like_count", "-view_count"]
            default: "-published_at"
        - name: published_after
          in: query
          schema:
            type: string
            format: date
        - name: published_before
          in: query
          schema:
            type: string
            format: date
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Creator's posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Pass your key as a Bearer token.
        Production keys: sk_live_*
        Sandbox keys: sk_test_*
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: string
        creator_id:
          type: string
        platform:
          type: string
        platform_post_id:
          type: string
        type:
          type: string
        content:
          type: object
          properties:
            text:
              type: string
            hashtags:
              type: array
              items:
                type: string
            mentions:
              type: array
              items:
                type: string
            media:
              type: array
              items:
                $ref: "#/components/schemas/Media"
        metrics:
          $ref: "#/components/schemas/PostMetrics"
        published_at:
          type: string
          format: date-time
        collected_at:
          type: string
          format: date-time
        metrics_updated_at:
          type: string
          format: date-time
    Media:
      type: object
      properties:
        type:
          type: string
          enum: [image, video]
        url:
          type: string
        thumbnail_url:
          type: string
        duration_seconds:
          type: number
        width:
          type: integer
        height:
          type: integer
    PostMetrics:
      type: object
      properties:
        like_count:
          type: integer
        comment_count:
          type: integer
        share_count:
          type: integer
        view_count:
          type: integer
        save_count:
          type: integer
        engagement_rate:
          type: number
...
    Pagination:
      type: object
      properties:
        has_more:
          type: boolean
        next_cursor:
          type: string
        total_count:
          type: integer
...
    PaginatedPosts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        pagination:
          $ref: "#/components/schemas/Pagination"
...